[PP-148] Text Rendering #2

+++ Project-Pentagram/Engine/Scene/MenuScene.h
@@ -4,13 +4,14 @@
 #include "Engine/SceneManager.h"
 #include "Engine/Input.h"
 #include "Engine/Collector/EngineDataCollector.h"
+#include "Utilities/Color.h"
 
 class MenuScene : public GameScene
 {
 	public:
 		virtual void GameSceneLoad() override;
 		virtual void GameSceneInit() override;
-		virtual void GameSceneUpdate(double dt) override;
+		virtual void GameSceneUpdate(float dt) override;
 		virtual void GameSceneDraw() override;
 		virtual void GameSceneUnload() override;
 		virtual void GameSceneFree() override;
