+++ Project-Pentagram/Engine/Scene/NodeScene.h
@@ -1,13 +1,15 @@
 #pragma once
 
 #include "Engine/GameScene.h"
+#include "Engine/SceneManager.h"
+#include "Engine/Input.h"
 
 class NodeScene : public GameScene
 {
 	public:
 		virtual void GameSceneLoad() override;
 		virtual void GameSceneInit() override;
-		virtual void GameSceneUpdate(double dt) override;
+		virtual void GameSceneUpdate(float dt) override;
 		virtual void GameSceneDraw() override;
 		virtual void GameSceneUnload() override;
 		virtual void GameSceneFree() override;
